feat: extract platformList to common package
